# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tasteful food ordering API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: string;
  name: string;
  slug: string;
  description: string;
  emoji: string;
  dishCount: number;
}

export interface Dish {
  id: string;
  categoryId: string;
  categoryName: string;
  name: string;
  tagline: string;
  description: string;
  price: number;
  imageUrl: string;
  rating: number;
  ratingCount: number;
  prepTimeMinutes: number;
  calories: number;
  /** 0-3 chili peppers */
  spiceLevel: number;
  isVegetarian: boolean;
  isPopular: boolean;
  isFeatured: boolean;
  ingredients: string[];
}

export interface OrderItem {
  dishId: string;
  dishName: string;
  unitPrice: number;
  quantity: number;
  imageUrl: string;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  received: "received",
  preparing: "preparing",
  on_the_way: "on_the_way",
  delivered: "delivered",
} as const;

export interface Order {
  id: string;
  orderNumber: string;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  deliveryAddress: string;
  notes: string;
  status: OrderStatus;
  subtotal: number;
  deliveryFee: number;
  tax: number;
  total: number;
  items: OrderItem[];
  createdAt: string;
  estimatedReadyAt: string;
}

export interface CreateOrderItemInput {
  dishId: string;
  /** @minimum 1 */
  quantity: number;
}

export interface CreateOrderInput {
  /** @minLength 1 */
  customerName: string;
  customerEmail: string;
  /** @minLength 1 */
  customerPhone: string;
  /** @minLength 1 */
  deliveryAddress: string;
  notes?: string;
  /** @minItems 1 */
  items: CreateOrderItemInput[];
}

export interface StatsSummary {
  totalDishes: number;
  totalCategories: number;
  ordersToday: number;
  averageRating: number;
  averagePrepMinutes: number;
}

export interface PopularDish {
  dishId: string;
  name: string;
  imageUrl: string;
  ordersCount: number;
  rating: number;
}

export type ListDishesParams = {
  categoryId?: string;
  search?: string;
  featured?: boolean;
};