# lib/api-zod/src/generated/types/category.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tasteful food ordering API
 * OpenAPI spec version: 0.1.0
 */

export interface Category {
  id: string;
  name: string;
  slug: string;
  description: string;
  emoji: string;
  dishCount: number;
}