# lib/api-zod/src/generated/types/createOrderInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tasteful food ordering API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderItemInput } from "./createOrderItemInput";

export interface CreateOrderInput {
  /** @minLength 1 */
  customerName: string;
  customerEmail: string;
  /** @minLength 1 */
  customerPhone: string;
  /** @minLength 1 */
  deliveryAddress: string;
  notes?: string;
  /** @minItems 1 */
  items: CreateOrderItemInput[];
}