# lib/api-zod/src/generated/types/createOrderItemInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tasteful food ordering API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateOrderItemInput {
  dishId: string;
  /** @minimum 1 */
  quantity: number;
}