# lib/api-zod/src/generated/types/dish.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tasteful food ordering API
 * OpenAPI spec version: 0.1.0
 */

export interface Dish {
  id: string;
  categoryId: string;
  categoryName: string;
  name: string;
  tagline: string;
  description: string;
  price: number;
  imageUrl: string;
  rating: number;
  ratingCount: number;
  prepTimeMinutes: number;
  calories: number;
  /** 0-3 chili peppers */
  spiceLevel: number;
  isVegetarian: boolean;
  isPopular: boolean;
  isFeatured: boolean;
  ingredients: string[];
}