# lib/api-zod/src/generated/types/popularDish.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tasteful food ordering API
 * OpenAPI spec version: 0.1.0
 */

export interface PopularDish {
  dishId: string;
  name: string;
  imageUrl: string;
  ordersCount: number;
  rating: number;
}