# lib/api-zod/src/generated/types/statsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tasteful food ordering API
 * OpenAPI spec version: 0.1.0
 */

export interface StatsSummary {
  totalDishes: number;
  totalCategories: number;
  ordersToday: number;
  averageRating: number;
  averagePrepMinutes: number;
}